+++ src/background/messenger/strict/api.ts
@@ -73,3 +73,10 @@ export const fetchFeatureFlagsInBackground = getMethod(
   "FETCH_FEATURE_FLAGS",
   bg,
 );
+
+export const services = {
+  locateAllForId: getMethod("LOCATE_SERVICES_FOR_ID", bg),
+  locate: getMethod("LOCATE_SERVICE", bg),
+  refresh: getMethod("REFRESH_SERVICES", bg),
+  refreshLocal: getMethod("LOCATOR_REFRESH_LOCAL", bg),
+};
